+++ app/src/main/java/com/jiangyy/wanandroid/logic/TodoUrl.kt
@@ -8,8 +8,9 @@ import rxhttp.wrapper.param.RxHttp
 class TodoUrl {
 
     companion object {
-        suspend fun pageTodo(page: Int): Await<Beans<Todo>> {
+        suspend fun pageTodo(page: Int, status: Int?): Await<Beans<Todo>> {
             return RxHttp.get("lg/todo/v2/list/$page/json")
+                .addQuery("status", status)
                 .toAwait()
         }
