+++ app/src/main/java/com/jiangyy/wanandroid/ui/todo/TodosActivity.kt
@@ -2,97 +2,33 @@ package com.jiangyy.wanandroid.ui.todo
 
 import android.content.Context
 import android.content.Intent
-import androidx.lifecycle.lifecycleScope
-import com.jiangyy.viewbinding.MultipleStateModule
+import androidx.fragment.app.Fragment
+import androidx.viewpager2.adapter.FragmentStateAdapter
 import com.jiangyy.viewbinding.base.BaseLoadActivity
 import com.jiangyy.wanandroid.databinding.ActivityTodosBinding
-import com.jiangyy.wanandroid.logic.TodoUrl
-import com.jiangyy.wanandroid.ui.adapter.TodoAdapter
-import kotlinx.coroutines.launch
-import rxhttp.awaitResult
 
-class TodosActivity : BaseLoadActivity<ActivityTodosBinding>(), MultipleStateModule {
-
-    private val mAdapter = TodoAdapter()
-    private var mPage = 1
+class TodosActivity : BaseLoadActivity<ActivityTodosBinding>() {
 
     override fun initValue() {
 
     }
 
     override fun initWidget() {
-        binding.recyclerView.adapter = mAdapter
-        binding.refreshLayout.setOnRefreshListener {
-            refresh()
+
+        binding.viewPager.adapter = object : FragmentStateAdapter(this) {
+            override fun getItemCount(): Int {
+                return 2
+            }
+
+            override fun createFragment(position: Int): Fragment {
+                return TodosFragment.newInstance(position)
+            }
         }
-        mAdapter.loadMoreModule.setOnLoadMoreListener {
-            loadMore()
-        }
+
     }
 
     override fun preLoad() {
-        refresh()
-    }
 
-    private fun refresh() {
-        mPage = 0
-        mAdapter.setList(null)
-        lifecycleScope.launch {
-            TodoUrl.pageTodo(mPage)
-                .awaitResult {
-                    binding.refreshLayout.isRefreshing = false
-                    if (it.isSuccess()) {
-                        if (it.data?.datas.isNullOrEmpty()) {
-                            preLoadWithEmpty("暂无数据")
-                        } else {
-                            preLoadSuccess()
-                            mAdapter.addData(it.data?.datas!!)
-                            if (mAdapter.data.size == it.data.total) {
-                                mAdapter.loadMoreModule.loadMoreEnd()
-                            } else {
-                                mAdapter.loadMoreModule.loadMoreComplete()
-                                ++mPage
-                            }
-                        }
-                    } else {
-                        preLoadWithFailure(it.errorMsg.orEmpty()) {
-                            preLoad()
-                        }
-                    }
-                }
-                .onFailure {
-                    binding.refreshLayout.isRefreshing = false
-                    preLoadWithFailure {
-                        preLoad()
-                    }
-                }
-        }
-    }
-
-    private fun loadMore() {
-        lifecycleScope.launch {
-            TodoUrl.pageTodo(mPage)
-                .awaitResult {
-                    if (it.isSuccess()) {
-                        if (it.data?.datas.isNullOrEmpty()) {
-                            mAdapter.loadMoreModule.loadMoreEnd()
-                        } else {
-                            mAdapter.addData(it.data?.datas!!)
-                            if (mAdapter.data.size == it.data.total) {
-                                mAdapter.loadMoreModule.loadMoreEnd()
-                            } else {
-                                mAdapter.loadMoreModule.loadMoreComplete()
-                                ++mPage
-                            }
-                        }
-                    } else {
-                        mAdapter.loadMoreModule.loadMoreFail()
-                    }
-                }
-                .onFailure {
-                    mAdapter.loadMoreModule.loadMoreFail()
-                }
-        }
     }
 
     companion object {
