+++ app/src/main/res/values/colors.xml
@@ -6,7 +6,7 @@
     <color name="colorOnPrimary">#FFFFFFFF</color>
     <color name="colorOnSecondary">#FF000000</color>
     <color name="colorCommon">#1F5B89</color>
-    <color name="colorBackground">#f8f8f8</color>
+    <color name="colorBackground">#F1F1F1</color>
     <color name="colorText">#FF000000</color>
     <color name="colorSubText">#8A000000</color>
     <color name="colorTag">#FFFFFF</color>
