+++ app/src/main/res/values/themes.xml
@@ -1,6 +1,6 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <style name="Theme.Wanandroid" parent="Theme.MaterialComponents.DayNight.NoActionBar">
-        <item name="android:colorBackground">#F1F1F1</item>
+        <item name="android:colorBackground">@color/colorBackground</item>
         <!-- Primary brand color. -->
         <item name="colorPrimary">@color/colorCommon</item>
         <item name="colorPrimaryVariant">@color/colorCommon</item>
